# public/api/old_swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "Modasti Backend API",
    "termsOfService": "",
    "contact": {
      "email": "biz@riseapps.co"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "stage-api.modasti.net",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/homeGroups": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "get home groups",
        "description": "",
        "operationId": "homeGroups",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Home Group",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "offset": {
                  "type": "number"
                },
                "limit": {
                  "type": "number"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "mine": {
                  "$ref": "#/definitions/Groups"
                },
                "in": {
                  "$ref": "#/definitions/Groups"
                },
                "other": {
                  "$ref": "#/definitions/Groups"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/groupDetails": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "group details",
        "description": "",
        "operationId": "groupDetails",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Home Group",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "groupId": {
                  "type": "number",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "group": {
                  "type": "object",
                  "$ref": "#/definitions/Group"
                },
                "photos": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "photo_id": {
                        "type": "number"
                      },
                      "photo": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "photo_name": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/createGroup": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "create group",
        "description": "",
        "operationId": "createGroup",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Create Group",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "required": true
                },
                "description": {
                  "type": "string"
                },
                "isPrivate": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Group"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/deleteGroup": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "delete group",
        "description": "",
        "operationId": "deleteGroup",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Delete Group",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "groupId": {
                  "type": "string",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/getGroupMembers": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "group members",
        "description": "",
        "operationId": "getGroupMembers",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Group Members",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "groupId": {
                  "type": "number",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/inviteMembers": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "invite members",
        "description": "",
        "operationId": "inviteMembers",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Invite Members To Group",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "groupId": {
                  "type": "number",
                  "required": true
                },
                "users": {
                  "type": "array",
                  "required": true,
                  "items": {
                    "type": "number"
                  }
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/getUserGroups": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "get User Groups",
        "description": "",
        "operationId": "getUserGroups",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User Groups",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "number",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "items": {
                  "$ref": "#/definitions/Groups"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/joinToGroup": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "join to group",
        "description": "",
        "operationId": "joinToGroup",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Join to Group",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "groupId": {
                  "type": "string",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "string",
                  "enum": [
                    "joined",
                    "invited"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/getJoinInvites": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "get join invites",
        "description": "",
        "operationId": "getJoinInvites",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "invited_id": {
                    "type": "number"
                  },
                  "group_id": {
                    "type": "number"
                  },
                  "user": {
                    "$ref": "#/definitions/User"
                  },
                  "group": {
                    "$ref": "#/definitions/Group"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/approveJoinInvite": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "approve join invite",
        "description": "",
        "operationId": "approveJoinInvite",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Approve Join Invite",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "groupId": {
                  "type": "number",
                  "required": true
                },
                "userId": {
                  "type": "number",
                  "required": true
                },
                "approved": {
                  "type": "boolean",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/leaveGroup": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "leave group",
        "description": "",
        "operationId": "leaveGroup",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "leaveGroup",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "groupId": {
                  "type": "number",
                  "required": "true"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_token",
      "in": "query"
    }
  },
  "definitions": {
...
    "Groups": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Group"
      }
    },
    "Group": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "is_private": {
          "type": "boolean"
        },
        "members_counter": {
          "type": "integer"
        },
        "image": {
          "type": "string"
        },
        "contests_counter": {
          "type": "integer"
        },
        "is_member": {
          "type": "boolean"
        },
        "photo_counter": {
          "type": "number"
        },
        "request_counter": {
          "type": "number"
        },
        "wins_counter": {
          "type": "number"
        },
        "is_invited": {
          "type": "boolean"
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "fname": {
          "type": "string"
        },
        "lname": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "user_type": {
          "type": "string"
        },
        "following_counter": {
          "type": "number"
        },
        "follower_counter": {
          "type": "number"
        },
        "is_followed": {
          "type": "boolean"
        },
        "photo": {
          "type": "object",
          "properties": {
            "table_id": {
              "type": "integer"
            },
            "photo_name": {
              "type": "string"
            }
          }
        }
      }
    },
...
  }
}